\name{gchol-methods}
\docType{methods}
\alias{gchol-methods}
\alias{gchol,matrix-method}
\title{ Methods for Function gchol in Package `kinship'}
\description{gchol methods, included from archived \code{kinship} package that can be found at http://cran.r-project.org/web/packages/kinship/index.html. }
\section{Methods}{\describe{

\item{x = "matrix"}{ an ordinary matrix or bdsmatrix object }
}}
\keyword{methods}
