\name{GWAF-package}
\alias{GWAF-package}
\alias{GWAF}
\docType{package}
\title{
Genome-wide association/interaction analysis and rare variant analysis with family data
}
\description{
For continuous traits, \code{GWAF} package provides two sets of functions for each of genome-wide association/interaction analyses 
with observed/imputed SNP genotypes for family data. One fits Linear Mixed Effects (LME) model and the other fits Generalized Estimation Equation (GEE) model 
to accounting for within pedigree correlation. While for dichotomous trait, \code{GWAF} package provides functions to fit GEE model for
genome-wide association/interaction analyses. For rare variant analysis, \code{GWAF} fits LME and generalized linear mixed effects (GLMM) model 
for continuous traits and dichotomous traits, respectively. In addition, \code{GWAF} package 
also provides functions for making genome-wide p-values plot and QQ plot that contains genomic control 
parameter estimate and generating scripts for genome-wide association analysis. 
Please read UsingGWAF.pdf for more information including examples and description to output files.
}
\details{
\tabular{ll}{
Package: \tab GWAF\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2013-06-06\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Qiong Yang <qyang@bu.edu> and Ming-Huei Chen <mhchen@bu.edu>

Maintainer: Ming-Huei Chen <mhchen@bu.edu>
}
