\name{lmekin}
\alias{lmekin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ function for linear mixed effects modeling with a kinship coefficient matrix }
\description{
   A similar function to \code{lmekin} from \code{kinship} package, but using Wald test, instead of t test. As \code{kinship} package has been archived, \code{GWAF} package now
includes functions from \code{kinship} package to support the modified \code{lmekin} function. Archived \code{kinship} package is available 
at http://cran.r-project.org/web/packages/kinship/index.html. 
}

