\name{GWAF-package}
\alias{GWAF-package}
\alias{GWAF}
\docType{package}
\title{
Genome-wide association/interaction analyses with family data
}
\description{
For continuous traits, \code{GWAF} package provides two sets of functions for each of genome-wide association analyses and genome-wide interaction analyses 
with observed/imputed SNP genotypes for family data. One fits Linear Mixed Effects (LME) model and the other fits Generalized Estimation Equation (GEE) model 
to accounting for within pedigree correlation. While for dichotomous trait, \code{GWAF} package provides functions to fit GEE model for
genome-wide association analyses and genome-wide interaction analyses. For LME, previous version of \code{GWAF} depends on \code{kinship} package that has been archived.
In the current version, \code{GWAF} includes functions from \code{kinship}. In addition, \code{GWAF} package 
also provides functions for making genome-wide p-values plot and QQ plot that contains genomic control 
parameter estimate and generating scripts for genome-wide association analysis.
}
\details{
\tabular{ll}{
Package: \tab GWAF\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2012-02-27\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Qiong Yang <qyang@bu.edu> and Ming-Huei Chen <mhchen@bu.edu>

Maintainer: Ming-Huei Chen <mhchen@bu.edu>
}
