\name{GWAF-package}
\alias{GWAF-package}
\alias{GWAF}
\docType{package}
\title{
Genome-Wide Association analyses with Family data
}
\description{
Functions to fit Linear Mixed Effects (LME) model accounting for within pedigree correlation 
for testing genetic association between genotyped/imputed SNPs and a continuous trait and functions to 
fit Generalized Estimation Equation (GEE) model using each pedigree as a cluster for 
testing genetic association between SNPs and a dichotomous trait. In addition, GWAF package 
also provides functions for making genome-wide p-values plot and QQ plot containing genomic control 
parameter estimate. 
}
\details{
\tabular{ll}{
Package: \tab GWAF\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2009-09-22\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Qiong Yang <qyang@bu.edu> and Ming-Huei Chen <mhchen@bu.edu>

Maintainer: Ming-Huei Chen <mhchen@bu.edu>
}
