\name{lmekin}
\alias{lmekin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ function for linear mixed effects modeling with a kinship coefficient matrix }
\description{
   A similar function to lmekin from kinship package, but using Wald test, instead of t test. Please see http://cran.r-project.org/web/packages/kinship/kinship.pdf
for more information.
}

\seealso{ \code{lmekin()} function from package \code{kinship} }
